+++ src/frontend/src/features/participantTile/components/KeyboardShortcutHint.tsx
