+++ src/frontend/src/features/participantTile/components/MutedMicIndicator.tsx
