+++ src/frontend/src/features/participantTile/components/ParticipantName.tsx
