+++ src/frontend/src/features/participantTile/components/ParticipantPlaceholder.tsx
